Reorganise Tfl code as a module (potential extraction as a library)
